authentication done

+++ src/blogapp/views.py
@@ -6,11 +6,21 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.generics import ListCreateAPIView,RetrieveUpdateDestroyAPIView,ListAPIView,CreateAPIView,RetrieveAPIView
+from rest_framework.filters import OrderingFilter
+from django.core.cache import cache
+from rest_framework.exceptions import PermissionDenied
 
 
 class CreateAndGetAllBlogs(ListCreateAPIView):
-    queryset=Blog.objects.all()
-    serializer_class=BlogSerializer
+    queryset = Blog.objects.select_related()
+    serializer_class = BlogSerializer
+    filter_backends = [OrderingFilter]  # Enables sorting
+    ordering_fields = ['title', 'id', "create_at", "update_at"]  # Fields to allow sorting by
+    ordering = ['id']  # Default ordering
+
+    
+
+
 
   
 class GetUpdateDeleteBlog(RetrieveUpdateDestroyAPIView):
@@ -20,6 +30,8 @@ class GetUpdateDeleteBlog(RetrieveUpdateDestroyAPIView):
     def perform_update(self, serializer):
         # Retrieve the existing instance before updating
         blog = self.get_object()
+        if blog.creator != self.request.user:
+            raise PermissionDenied("You do not have permission to update this blog.")
         old_image = blog.image  # Get the old image instance
         
         # Save the new data
@@ -33,13 +45,17 @@ class GetUpdateDeleteBlog(RetrieveUpdateDestroyAPIView):
             print("No image change detected or no old image to delete.")  # Debug log
 
     def perform_destroy(self, instance):
+        if instance.creator != self.request.user:
+            raise PermissionDenied("You do not have permission to delete this blog.")
         if instance.image:
             instance.image.delete(save=False)
         instance.delete()
 
     
 
-
-
+class GetUserBlogs(ListAPIView):
+    serializer_class=BlogSerializer
+    def get_queryset(self):
+        return Blog.objects.filter(creator=self.request.user)
